Change IDs to bigint

+++ nhs/src/main/java/uk/ac/cf/spring/nhs/Diary/Repository/PhotoRepository.java
@@ -8,5 +8,5 @@ import uk.ac.cf.spring.nhs.Diary.Entity.Photo;
 import java.util.List;
 
 public interface PhotoRepository extends JpaRepository<Photo, Integer> {
-    List<Photo> findByUserId(int userId, Sort date);
+    List<Photo> findByUserId(long userId, Sort date);
 }
